requirements.txt
autopep8==1.5.4
fastapi==0.61.1
flake8==3.8.4
google-cloud-firestore==1.9.0
gunicorn==20.0.4
httptools==0.1.1
mock-firestore==0.7.1
pytest-cov==2.7.1
pytest==4.3.1
python-language-server==0.30.0
requests==2.24.0
uvicorn==0.12.1
uvloop==0.14.0